# frontend/src/contracts/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MilestoneCreateRequest": {
        "description": "Request body for creating a milestone.",
        "properties": {
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "project_id": {
            "exclusiveMinimum": 0.0,
            "title": "Project Id",
            "type": "integer"
          },
          "sort_order": {
            "default": 0,
            "title": "Sort Order",
            "type": "integer"
          },
          "status": {
            "default": "backlog",
            "enum": [
              "backlog",
              "planned",
              "doing",
              "done",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "target_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          }
        },
        "required": [
          "project_id",
          "description"
        ],
        "title": "MilestoneCreateRequest",
        "type": "object"
      },
      "MilestoneListResponse": {
        "description": "List of milestones.",
        "properties": {
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/MilestoneResponse"
            },
            "title": "Milestones",
            "type": "array"
          }
        },
        "required": [
          "milestones"
        ],
        "title": "MilestoneListResponse",
        "type": "object"
      },
      "MilestoneResponse": {
        "description": "Milestone response body, including associated session minutes.",
        "properties": {
          "completed_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Date"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "notes": {
            "title": "Notes",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "integer"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          },
          "status": {
            "enum": [
              "backlog",
              "planned",
              "doing",
              "done",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "target_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          },
          "total_session_minutes": {
            "title": "Total Session Minutes",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "description",
          "status",
          "completed_date",
          "target_date",
          "sort_order",
          "notes",
          "total_session_minutes",
          "created_at",
          "updated_at"
        ],
        "title": "MilestoneResponse",
        "type": "object"
      },
      "MilestoneStatusRequest": {
        "description": "Request body for a milestone status transition.",
        "properties": {
          "status": {
            "enum": [
              "backlog",
              "planned",
              "doing",
              "done",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "MilestoneStatusRequest",
        "type": "object"
      },
      "MilestoneUpdateRequest": {
        "description": "Request body for updating a milestone.",
        "properties": {
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "sort_order": {
            "default": 0,
            "title": "Sort Order",
            "type": "integer"
          },
          "status": {
            "enum": [
              "backlog",
              "planned",
              "doing",
              "done",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "target_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          }
        },
        "required": [
          "description",
          "status"
        ],
        "title": "MilestoneUpdateRequest",
        "type": "object"
      },
      "OkResponse": {
        "description": "Generic success acknowledgement for operations without a body.\n\n:param ok: Always ``True``.",
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Portfolio Manager Sidecar",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/milestones": {
      "post": {
        "description": "Create a milestone.",
        "operationId": "create_milestone_api_v1_milestones_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MilestoneCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Milestone",
        "tags": [
          "milestones"
        ]
      }
    },
    "/api/v1/milestones/{milestone_id}": {
      "delete": {
        "description": "Delete a milestone permanently.",
        "operationId": "delete_milestone_api_v1_milestones__milestone_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "title": "Milestone Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Milestone",
        "tags": [
          "milestones"
        ]
      },
      "get": {
        "description": "Fetch a single milestone.",
        "operationId": "get_milestone_api_v1_milestones__milestone_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "title": "Milestone Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Milestone",
        "tags": [
          "milestones"
        ]
      },
      "put": {
        "description": "Update a milestone's editable fields.",
        "operationId": "update_milestone_api_v1_milestones__milestone_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "title": "Milestone Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MilestoneUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Milestone",
        "tags": [
          "milestones"
        ]
      }
    },
    "/api/v1/milestones/{milestone_id}/status": {
      "post": {
        "description": "Transition a milestone to a new status.",
        "operationId": "set_status_api_v1_milestones__milestone_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "title": "Milestone Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MilestoneStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Status",
        "tags": [
          "milestones"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/milestones": {
      "get": {
        "description": "List a project's milestones (ordered) with session-minute totals.",
        "operationId": "list_milestones_api_v1_projects__project_id__milestones_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Milestones",
        "tags": [
          "milestones"
        ]
      }
    },
...
  }
}